Fix rrand warning in main.c
On Linux, it produced this warning:

integer overflow in expression of type ‘int’ results in ‘-2147483648’

This was presumably because RAND_MAX is the same as the maximum integer value,
so RAND_MAX+1 overflowed.

The fix is not perfect as using a simple modulus will make some values in the
range slightly more likely than others. But the probability difference is
relative to the size of the range and it is only ever used on a small range (a
quick search of the code base finds 2, 30 and 256), and it doesn't seem to be
used for anything critical anyway.